Rename Account to MainApp
