+++ MainApp/admin.py
