+++ MainApp/templates/MainApp/base_profile.html
@@ -331,10 +331,10 @@
         <a href="#">Refresh</a>
         <a href="#">Edit</a>
         {% if validSession %}
-        <a href="{% url 'Account:logout' %}" style="float:right">Logout</a>
+        <a href="{% url 'MainApp:logout' %}" style="float:right">Logout</a>
         {% else %}
-        <a href="{% url 'Account:register' %}" style="float:right">Register</a>
-        <a href="{% url 'Account:login' %}" style="float:right">Login</a>
+        <a href="{% url 'MainApp:register' %}" style="float:right">Register</a>
+        <a href="{% url 'MainApp:login' %}" style="float:right">Login</a>
 
          {% endif %}
       </div>
