+++ MainApp/templates/MainApp/homepage.html
@@ -1,6 +1,6 @@
 
 
-{% extends "Account/base_homepage.html" %}
+{% extends "MainApp/base_homepage.html" %}
 
 <!-- Navbar -->
 {% block navbar %}
@@ -13,10 +13,10 @@
   <a href=".">Refresh</a>  
 
   {% if validSession %}
-  <a href="{% url 'Account:logout' %}" style="float:right">Logout</a>
+  <a href="{% url 'MainApp:logout' %}" style="float:right">Logout</a>
   {% else %}
-  <a href="{% url 'Account:register' %}" style="float:right">Register</a>
-  <a href="{% url 'Account:login' %}" style="float:right">Login</a>
+  <a href="{% url 'MainApp:register' %}" style="float:right">Register</a>
+  <a href="{% url 'MainApp:login' %}" style="float:right">Login</a>
   <!--Commenting out the modal button for now -->
   <!--<button style="float: right;">Login</button>--><!-- https://www.w3schools.com/howto/howto_css_modals.asp Code for making a modal popup window -->
 
@@ -39,7 +39,7 @@
 <!-- Card that holds Profile information -->
 {% block Profile_Card %}
   {% if validSession %}
-  <a href="{% url 'Account:profile' user %}" style="text-decoration: none;
+  <a href="{% url 'MainApp:profile' user %}" style="text-decoration: none;
   outline: none;">
     <div class="card">
       <div class="fakeimg" style="width: 30%; float: left;">Profile pic.</div>
@@ -67,7 +67,7 @@
             <h3 style="color: rgb(99, 99, 99);">@{{user.username}}</h3>
           </a>
           <div style="width: 30%; float: right;"> 
-            <form method="post" action="{% url 'Account:profile' user.username %}">
+            <form method="post" action="{% url 'MainApp:profile' user.username %}">
               {% csrf_token %}
                 {% if not alreadyFollow %}
                   <button name="follow" type="submit" style="float:right" value={{user.username}} class="follow_button">Follow</button>
