+++ MainApp/templates/MainApp/profile.html
@@ -1,5 +1,5 @@
 
-{% extends "Account/base_profile.html" %}
+{% extends "MainApp/base_profile.html" %}
 
 {% block title %}
 <title>{{profile_user.first_name}} {{profile_user.last_name}} (@{{profile_user.username}})</title>
@@ -17,7 +17,7 @@
         <h1>{{profile_user.first_name}} {{profile_user.last_name}} </h1>
         <p>{{profile_user.username}}</p>
       </div>
-        <form method="post" action="{% url 'Account:profile' profile_user.username %}">
+        <form method="post" action="{% url 'MainApp:profile' profile_user.username %}">
           {% csrf_token %}
           {% if not isNative %}
             {% if not auth_follow %}
@@ -32,7 +32,7 @@
       <h1>Bio</h1>
       <p>Birthdate</p>
   </div>
-    {% endblock %}
+{% endblock %}
 
 {% block Personal_Scroll %}
 
@@ -73,7 +73,7 @@
             <h3 style="color: rgb(99, 99, 99);">@{{user.username}}</h3>
           </a>
           <div style="width: 30%; float: right;"> 
-            <form method="post" action="{% url 'Account:profile' user.username %}">
+            <form method="post" action="{% url 'MainApp:profile' user.username %}">
               {% csrf_token %}
                 {% if not alreadyFollow %}
                   <button name="follow" type="submit" style="float:right" value={{user.username}} class="follow_button">Follow</button>
@@ -96,7 +96,7 @@
 {% block navbar %}
 
       <div class="topnav">
-        <a href="{% url 'Account:index' %}">Home</a>
+        <a href="{% url 'MainApp:index' %}">Home</a>
         <a href="#">Followers {{followed_by_len}}</a>
         <a href="#">Following {{following_len}}</a>
         <a href="#">Likes</a>
@@ -104,10 +104,10 @@
         <a href="#">Edit</a>
 
         {% if validSession %}
-        <a href="{% url 'Account:logout' %}" style="float:right">Logout</a>
+        <a href="{% url 'MainApp:logout' %}" style="float:right">Logout</a>
         {% else %}
-        <a href="{% url 'Account:register' %}" style="float:right">Register</a>
-        <a href="{% url 'Account:login' %}" style="float:right">Login</a>
+        <a href="{% url 'MainApp:register' %}" style="float:right">Register</a>
+        <a href="{% url 'MainApp:login' %}" style="float:right">Login</a>
 
   {% endif %}
       </div>
