+++ MainApp/templates/registration/logout.html
