+++ MainApp/templates/registration/register.html
@@ -12,7 +12,7 @@
 
 
 
-<form method="post" action="{% url 'Account:register' %}">
+<form method="post" action="{% url 'MainApp:register' %}">
 
     {% csrf_token %}
