+++ MainApp/templatetags/nbsp.py
