+++ MainApp/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from . import views
 from django.contrib.auth import views as auth_views
 
-app_name = 'Account'
+app_name = 'MainApp'
 urlpatterns = [
     path('',views.index,name='index'), 
     path('register/',views.register,name='register'),
